Incremental update working
