--- gnocchi-cli/src/main/scala/net/fnothaft/gnocchi/cli/LoadGnocchiModel.scala
@@ -1,37 +1,0 @@
-///**
-// * Licensed to Big Data Genomics (BDG) under one
-// * or more contributor license agreements.  See the NOTICE file
-// * distributed with this work for additional information
-// * regarding copyright ownership.  The BDG licenses this file
-// * to you under the Apache License, Version 2.0 (the
-// * "License"); you may not use this file except in compliance
-// * with the License.  You may obtain a copy of the License at
-// *
-// *     http://www.apache.org/licenses/LICENSE-2.0
-// *
-// * Unless required by applicable law or agreed to in writing, software
-// * distributed under the License is distributed on an "AS IS" BASIS,
-// * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// * See the License for the specific language governing permissions and
-// * limitations under the License.
-// */
-//package net.fnothaft.gnocchi.cli
-//
-//import java.io.{ FileInputStream, ObjectInputStream }
-//
-//import net.fnothaft.gnocchi.models.GnocchiModel
-//
-///**
-// * Loads a GnocchiModel from disk
-// */
-//object LoadGnocchiModel {
-//  //TODO: change loading to Spark Serialization
-//  def apply(readFrom: String): GnocchiModel = {
-//    val fis = new FileInputStream(readFrom)
-//    val ois = new ObjectInputStream(fis)
-//
-//    val model = ois.readObject.asInstanceOf[GnocchiModel]
-//    ois.close()
-//    model
-//  }
-//}
