--- gnocchi-cli/src/main/scala/net/fnothaft/gnocchi/cli/SaveGnocchiModel.scala
@@ -1,32 +1,0 @@
-///**
-// * Licensed to Big Data Genomics (BDG) under one
-// * or more contributor license agreements.  See the NOTICE file
-// * distributed with this work for additional information
-// * regarding copyright ownership.  The BDG licenses this file
-// * to you under the Apache License, Version 2.0 (the
-// * "License"); you may not use this file except in compliance
-// * with the License.  You may obtain a copy of the License at
-// *
-// *     http://www.apache.org/licenses/LICENSE-2.0
-// *
-// * Unless required by applicable law or agreed to in writing, software
-// * distributed under the License is distributed on an "AS IS" BASIS,
-// * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// * See the License for the specific language governing permissions and
-// * limitations under the License.
-// */
-//package net.fnothaft.gnocchi.cli
-//
-//import java.io.{ FileOutputStream, ObjectOutputStream }
-//import net.fnothaft.gnocchi.models.GnocchiModel
-//
-//object SaveGnocchiModel {
-//  //
-//  //  def apply(model: GnocchiModel, saveTo: String): Unit = {
-//  //
-//  //    val oos = new ObjectOutputStream(new FileOutputStream(saveTo))
-//  //    oos.writeObject(model)
-//  //    oos.close()
-//  //
-//  //  }
-//}
