+++ gnocchi-core/pom.xml
@@ -102,5 +102,20 @@
       <groupId>org.apache.spark</groupId>
       <artifactId>spark-mllib_${scala.version.prefix}</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
+      <version>1.7.5</version>
+    </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-simple</artifactId>
+      <version>1.6.4</version>
+    </dependency>
+    <dependency>
+      <groupId>org.scala-lang.modules</groupId>
+      <artifactId>scala-pickling_2.11</artifactId>
+      <version>0.10.1</version>
+    </dependency>
   </dependencies>
 </project>
