+++ pom.xml
@@ -392,16 +392,6 @@
         <version>${spark.version}</version>
         <scope>provided</scope>
       </dependency>
-      <dependency>
-       <groupId>org.slf4j</groupId>
-       <artifactId>slf4j-api</artifactId>
-       <version>1.7.5</version>
-      </dependency>
-      <dependency>
-        <groupId>org.slf4j</groupId>
-        <artifactId>slf4j-simple</artifactId>
-        <version>1.6.4</version>
-      </dependency>
     </dependencies>
   </dependencyManagement>
